requirements.txt
numpy>=1.21.0
pandas>=1.3.0
torch>=1.10.0
torchvision>=0.11.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
pyyaml>=6.0
argparse>=1.4.0
